Refactored some more components
