+++ packages/attendance/tsconfig.json
@@ -2,10 +2,12 @@
   "include": ["src"],
   "exclude": ["node_modules", "*/**/*.json"],
   "compilerOptions": {
+    "noImplicitAny": false,
     "jsx": "react",
+    "allowJs": true,
     "resolveJsonModule": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
-    "baseUrl": "src"
+    "baseUrl": "src",
   }
 }
